server/requirements.txt
pycrypto
django
paho-mqtt
django-extensions
gunicorn
psycopg2-binary
djangorestframework
django-mptt
django-suit
requests
django-model-utils
wechatpy
cryptography
djangorestframework_simplejwt